backend/puzzle/serializers/inventory.py
from rest_framework import serializers
from puzzle.models import Inventory


class InventorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Inventory
        fields = "__all__"
