requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
mysqlclient==2.0.3
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
